feat(quantities): sell products by sub-unit via a (units, sub) pair
Sale orders could not express a fraction of a product, and the Sbu* sub-unit
columns on Product were inert: the sale page pushed a sub-unit count straight
into DetailQuantities, so selling 3 candies deducted 3 boxes.

A quantity is now core.Quantity{Units, Sub} read at a divisor
(Product.SbuQuantity, int16, max 1000), worth Units + Sub/Divisor. Units is
never scaled, so there is no per-product "discrete or continuous" choice to make
at creation and get wrong later — grams are just divisor 1000.

Stored two ways, split by what the row is for:

  - Aggregates split. Stock, the movement ledger and the sale summaries keep
    Quantity and SubQuantity in separate columns. Sub stays unnormalized there,
    which is what makes addition closed without a carry: {a,b} + {c,d} =
    {a+c, b+d}. The stock engine keeps a plain integer +, and Scylla can still
    SUM() both columns.
  - Documents pack. A sale-order line stores Units*1000 + Sub in one column
    (1001 at divisor 6 = one box plus one candy) with its own DetailSubDivisor.
    A document is never accumulated, so it pays one column instead of two.

Divisors may only be refined to a multiple (6->12 yes, 12->6 no, since 9/12 is
4.5 sixths). Adding a sub-unit to a product that already has stock and sales
rewrites nothing: every existing row holds SubQuantity = 0, and zero sixths and
zero thousandths are the same zero. A stock row converts itself lazily when a
finer movement arrives, so no bulk ledger rewrite is ever needed.

Three pre-existing engine defects had to be fixed for it to be safe:

  - stock_movement_apply.go tested stock.Quantity < 0, which misses {0,-4} —
    four sub-units oversold, whose whole-unit half is exactly zero.
  - ProductStock.SelfParse derived Status from the whole-unit halves only, so a
    row holding 0 boxes and 4 loose candies was evicted from the delta sync.
  - MonetaryValue ignored the sub-unit part entirely.

PostSaleOrder now loads products to resolve the divisor, and re-reads both
prices from the catalog. Until now DetailPrices and TotalAmount came straight
off the request body unchecked, so a client could book a sale at any price.

At the SUNAT boundary a line with both halves becomes two document lines, priced
at FinalPrice and SbuFinalPrice. Every emitted quantity stays a whole count, so
no sixth of a box reaches the XML — facturago scales by 1e6, where 1/6 is not
representable.

Also in this commit, and unrelated to the above:

  - fn-init seeds each company its Site, Warehouse and CashBank, so a seeded
    company is operational. That removed the reason /initial-data existed as a
    standalone route, and with it hasPendingInitialData — two ScyllaDB queries
    on every login. The sign-up wizard already renders InitialDataForm, which
    stays.
  - Company 1 is exempt from credit budgets but not from permissions: credits
    zero out in chargeConfiguredCredits while the access frame still goes out.
    An exhausted budget was locking the operator out of its own software.
  - Carried along from the working tree: the Go 1.26 -> 1.27 toolchain bump,
    the genix-orm submodule pointer, docs/COLBIN_RUST_PLAN.md, .mcp.json and a
    menu label in frontend/core/modules.ts.

Design and rejected alternatives: docs/QUANTITY_SCALE_PLAN.md, plus RATIONALE
entries under backend/, backend/logistics/ and backend/sales/.

+++ backend/business/types/productos.go
@@ -37,23 +37,29 @@ const (
 
 type Product struct {
 	db.TableStruct[ProductTable, Product]
-	CompanyID     int32   `json:",omitempty"`
-	ID            int32   `db:"id,pk"`
-	TempID        int32   `json:",omitempty"`
-	Name          string  `db:"nombre"`
-	Description   string  `json:",omitempty"`
-	ContentHTML   string  `json:",omitempty"`
-	CategoryIDs   []int32 `json:",omitempty" db:"category_ids"`
-	BrandID       int32   `json:",omitempty"`
-	Params        []int8  `json:",omitempty"`
-	Price         int32   `json:",omitempty"`
-	CurrencyID    int16   `json:",omitempty"`
-	UnitID        int16   `json:",omitempty"`
-	Discount      float32 `json:",omitempty"`
-	FinalPrice    int32   `json:",omitempty"`
-	Weight        float32 `json:",omitempty"`
-	Volume        float32 `json:",omitempty"`
-	SbuQuantity   int32   `json:",omitempty"`
+	CompanyID   int32   `json:",omitempty"`
+	ID          int32   `db:"id,pk"`
+	TempID      int32   `json:",omitempty"`
+	Name        string  `db:"nombre"`
+	Description string  `json:",omitempty"`
+	ContentHTML string  `json:",omitempty"`
+	CategoryIDs []int32 `json:",omitempty" db:"category_ids"`
+	BrandID     int32   `json:",omitempty"`
+	Params      []int8  `json:",omitempty"`
+	Price       int32   `json:",omitempty"`
+	CurrencyID  int16   `json:",omitempty"`
+	UnitID      int16   `json:",omitempty"`
+	Discount    float32 `json:",omitempty"`
+	FinalPrice  int32   `json:",omitempty"`
+	Weight      float32 `json:",omitempty"`
+	Volume      float32 `json:",omitempty"`
+	// Sub-unit: the divisor of the (Units, Sub) quantity pair. SbuQuantity is how many
+	// sub-units make one whole unit — 1000 for a kilogram sold by the gram, 6 for a box sold
+	// by the candy. 0 or 1 means the product has no sub-unit. It may only ever be refined to
+	// a multiple (see core.IsQuantityDivisorRefinement): coarsening would round away real
+	// stock, since 9/12 is 4.5 sixths. int16 to match SubDivisor on the stock and movement
+	// rows, so one concept has one type.
+	SbuQuantity   int16   `json:",omitempty"`
 	SbuUnit       string  `json:",omitempty"`
 	SbuPrice      int32   `json:",omitempty"`
 	SbuDiscount   float32 `json:",omitempty"`
@@ -115,7 +121,7 @@ type ProductTable struct {
 	FinalPrice          db.Col[*ProductTable, int32]
 	Weight              db.Col[*ProductTable, float32]
 	Volume              db.Col[*ProductTable, float32]
-	SbuQuantity         db.Col[*ProductTable, int32]
+	SbuQuantity         db.Col[*ProductTable, int16]
 	SbuUnit             db.Col[*ProductTable, string]
 	SbuPrice            db.Col[*ProductTable, int32]
 	SbuDiscount         db.Col[*ProductTable, float32]
